// src/zerro-core/domain/zerro/hidden-data/types.ts
export enum HiddenDataType {
  Goals = 'goals',
  FxRates = 'fxRates',
  Budgets = 'budgets',
  LinkedAccounts = 'linkedAccounts',
  LinkedDebtors = 'linkedDebtors',
  EnvelopeMeta = 'EnvelopeMeta',
  UserSettings = 'UserSettings',
  TagOrder = 'tagOrder',
}

export type THiddenDataComment<TPayload = unknown> = {
  type: HiddenDataType
  month?: string
  payload: TPayload
}